Move `isRegisteredProperty()` to a separate file
